Extract Task::ReportingPeriod for getting due_date
This way it's easier to re-use the code elsewhere when we need to know
the due date on a reporting period, e.g. in
Task::AnticipatedUserNotificationList.
